% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{anonymize_dataframe}
\alias{anonymize_dataframe}
\title{this function transform the data to an SDCmicro object}
\usage{
anonymize_dataframe(
  sdc_data_frame,
  anony_methods = c("Suppression", "microaggregation"),
  k_anon_number = 2,
  aggregation_number = 3
)
}
\arguments{
\item{anony_methods}{vector of procedures names - "Suppression","microaggregation"}

\item{k_anon_number}{integer -  number of anonmity to achive}

\item{aggregation_number}{integer -  number of buckets for numeric feature}

\item{microdata}{sdc_data_frame sdc object or microdata}
}
\value{
anonymized SDC dataframe
}
\description{
this function transform the data to an SDCmicro object
}
