% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_low_sample_frequency}
\alias{detect_low_sample_frequency}
\title{this split a dataframe into lowrisk patients and highrisk patients}
\usage{
detect_low_sample_frequency(sdc_data_frame, k_anon_number = 2)
}
\arguments{
\item{k_anon_number}{integer -  number of anonmity to decide our split at}

\item{microdata}{sdc_data_frame sdc object or microdata}
}
\value{
list - two elements: 1) low_risk_patients 2)high_risk_patients
}
\description{
this split a dataframe into lowrisk patients and highrisk patients
}
