% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{safe_harbor_transformation}
\alias{safe_harbor_transformation}
\title{this function perform safe habor guidelines on the data}
\usage{
safe_harbor_transformation(microdata, unique_identifers, dates, dob, zipcode)
}
\arguments{
\item{microdata}{dataframe}

\item{unique_identifers}{vector of columns names -unique identifying number, characteristic, or code. Certificate/license numbers}

\item{dates}{vector of columns names -  admission date, discharge date, death date}

\item{dob}{columns name - birth date}

\item{zipcode}{columns name - geographic subdivisions}
}
\value{
anonymized data frame with safe harbor guidlines
}
\description{
this function perform safe habor guidelines on the data
}
