% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{transform_to_sdc}
\alias{transform_to_sdc}
\title{this function transform the data to an SDCmicro object}
\usage{
transform_to_sdc(microdata, keyVars, numVars)
}
\arguments{
\item{microdata}{dataframe}

\item{keyVars}{vector of columns names - are a set of variables that, when considered together, can be used to identify individual units.}

\item{numVars}{vector of columns names -  ontinuous key variables}

\item{pramVars}{vector of columns names -  Indices or names of categorical variables considered to be pramed}
}
\value{
sdcobject
}
\description{
this function transform the data to an SDCmicro object
}
